Moving NavigationBarModel to ViewModel
